patching building
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@2853 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/router/ValueTimeLine.h
@@ -405,7 +405,7 @@ public:
         ret = new T[myValues.size()];
         begin = myValues.begin()->first.first;
         end = (myValues.end()-1)->first.second;//begin + myValues.size() * interval;
-        for(int i=0; i<myValues.size(); i++) {
+        for(size_t i=0; i<myValues.size(); i++) {
             ret[i] = myValues[i].second;
         }
         return ret;
